requirements.txt
# Web & API
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
aiofiles==24.1.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# DB & Cache & Queue
SQLAlchemy==2.0.34
asyncpg==0.29.0
redis==5.0.7
celery==5.4.0
kombu==5.4.0

# Graph DB
neo4j==5.23.1

# Storage (MinIO for local dev)
minio==7.2.7

# NLP / Parsing (placeholders for MVP)
python-pptx==1.0.2
PyMuPDF==1.24.9

# Utils
tenacity==9.0.0
requests==2.32.3
python-dotenv==1.0.1
